rt: experiment with alt file structure
